Model search, UI enhancements, notifications implemented, need to finish managing collaborators and comparing models.

+++ src/main/java/com/modelbox/controllers/myModels/myModelsController.java
@@ -10,29 +10,36 @@ import javafx.fxml.FXMLLoader;
 import javafx.geometry.Pos;
 import javafx.scene.*;
 import javafx.scene.control.Button;
+import javafx.scene.control.CheckBox;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
 import javafx.scene.image.ImageView;
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyEvent;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.FlowPane;
+import javafx.scene.layout.HBox;
 import javafx.scene.layout.StackPane;
 import javafx.scene.shape.MeshView;
 import javafx.scene.shape.TriangleMesh;
+import javafx.scene.text.Text;
 import javafx.stage.FileChooser;
-import javafx.stage.Stage;
-import org.apache.commons.io.FilenameUtils;
 import org.bson.BsonDocument;
+
 import java.nio.file.Files;
 
 
 public class myModelsController {
 
+    @FXML public HBox myModelsToolBar;
     @FXML public ScrollPane myModelsScrollPane;
     @FXML public FlowPane myModelsFlowPane;
+    @FXML public Button noModelSearchResultsBtn;
     @FXML public Button noModelsBtn;
     @FXML public AnchorPane myModelsAnchorPane;
     @FXML public AnchorPane loadingAnchorPane;
     @FXML public TextField modelSearchField;
+    private int checkboxCount = 0;
 
     /**
      * Handles the search query specified by the user
@@ -40,12 +47,80 @@ public class myModelsController {
      * @param  event a JavaFX Event
      */
     @FXML
-    private void modelSearchEnterKeyPressed(Event event) {
+    private void searchModelBtnClicked(Event event) {
         try {
-
+            searchForModels();
         } catch (Exception exception) {
             exception.printStackTrace();
         }
+    }
+
+    /**
+     * Handles the search query specified by the user
+     *
+     * @param  event a JavaFX Event
+     */
+    @FXML
+    private void searchModelEnterKeyPressed(KeyEvent event) {
+        try {
+            if(event.getCode().equals((KeyCode.ENTER))) {
+                searchForModels();
+            }
+        } catch (Exception exception) {
+            exception.printStackTrace();
+        }
+    }
+
+    private void searchForModels() {
+        try {
+            myModelsToolBar.setVisible(false);
+            noModelsBtn.setVisible(false);
+            myModelsScrollPane.setVisible(false);
+            loadingAnchorPane.setVisible(true);
+
+            // Start the search
+            String functionCall = String.format("ModelBox.Models.getCurrentUserModelSearch('%s');", modelSearchField.getText());
+            app.mongoApp.eval(functionCall);
+        } catch (Exception exception) {
+            exception.printStackTrace();
+        }
+
+    }
+
+    /**
+     * Handles clearing the search query specified by the user and refreshing the view
+     *
+     * @param  event a JavaFX Event
+     */
+    @FXML
+    private void cancelSearchModelBtnClicked(Event event) {
+        try {
+            // Show my models view
+            app.viewLoader = new FXMLLoader(getClass().getResource("/views/myModels/myModels.fxml"));
+            Parent root = app.viewLoader.load();
+            app.myModelsView = app.viewLoader.getController();
+            app.dashboard.dashViewsAnchorPane.getChildren().setAll(root);
+
+            // Asynchronously populate the my models view and show appropriate nodes when ready
+            String functionCall = String.format("ModelBox.Models.getCurrentUserModels();");
+            app.mongoApp.eval(functionCall);
+        } catch (Exception exception) {
+            exception.printStackTrace();
+        }
+    }
+
+    /**
+     * Handles showing a checkbox on each model card to compare two selected models
+     *
+     * @param  event a JavaFX Event
+     */
+    @FXML
+    private void compareModelsBtnClicked(Event event) {
+        /*
+        for (int i = 0; i < myModelsFlowPane.getChildren().size(); i++) {
+            // For each model card activate (make visible) the checkbox to compare two models.
+            ((StackPane) myModelsFlowPane.getChildren().get(i))
+        }*/
     }
 
     /**
@@ -109,14 +184,11 @@ public class myModelsController {
             downloadModelBtn.setStyle("-fx-background-color: none;");
             downloadModelBtn.setOnAction(downloadModelBtnClicked);
 
-            // Create a share btn for each model card generated
-            Button shareModelBtn = new Button();
-            ImageView shareModelIcon = model.get("isShared").asBoolean().getValue() ? new ImageView("/images/multi-user-btn.png") : new ImageView("/images/share-btn.png");
-            shareModelIcon.setFitHeight(25);
-            shareModelIcon.setFitWidth(25);
-            shareModelBtn.setGraphic(shareModelIcon);
-            shareModelBtn.setStyle("-fx-background-color: none;");
-            shareModelBtn.setOnAction(shareModelBtnClicked);
+            // Create a hidden checkbox for each model card generated for use when comparing two models
+            CheckBox compareCheckbox = new CheckBox();
+            compareCheckbox.setVisible(false);
+            compareCheckbox.setStyle("-fx-background-color: transparent; -fx-border-color: #181a1d; -fx-border-radius: 5;");
+            compareCheckbox.setOnAction(compareCheckboxClicked);
 
             // Add the mesh to the model card group
             Group modelCardGroup = new Group(modelMeshView);
@@ -131,7 +203,7 @@ public class myModelsController {
             modelSubScene.setCamera(camera);
 
             // Manipulate the features of the model card and the arrangement of its internals
-            StackPane modelMeshPane = new StackPane(modelSubScene, deleteModelBtn, previewModelBtn, downloadModelBtn, shareModelBtn);
+            StackPane modelMeshPane = new StackPane(modelSubScene, deleteModelBtn, previewModelBtn, downloadModelBtn, compareCheckbox);
             modelMeshPane.setId(model.get("_id").asObjectId().getValue().toHexString());
             modelMeshPane.setStyle("-fx-background-color: #eeeeee; -fx-background-radius: 8 8 8 8");
             modelMeshPane.setMinWidth(150);
@@ -140,11 +212,57 @@ public class myModelsController {
             modelMeshPane.setMaxHeight(250);
             StackPane.setAlignment(modelSubScene, Pos.CENTER);
             StackPane.setAlignment(deleteModelBtn, Pos.TOP_RIGHT);
+            StackPane.setAlignment(compareCheckbox, Pos.BOTTOM_LEFT);
             StackPane.setAlignment(previewModelBtn, Pos.BOTTOM_RIGHT);
             previewModelBtn.setTranslateX(-30);
-            StackPane.setAlignment(shareModelBtn, Pos.BOTTOM_RIGHT);
-            shareModelBtn.setTranslateX(-60);
             StackPane.setAlignment(downloadModelBtn, Pos.BOTTOM_RIGHT);
+
+            if (model.get("owner_id").asString().getValue().equals(app.users.ownerId) && model.get("isShared").asBoolean().getValue()) {
+                // Current user is the owner and the model is shared
+
+                // Show multi-user btn on models that apply
+                Button shareModelBtn = new Button();
+                ImageView shareModelIcon = new ImageView("/images/multi-user-btn.png");
+                shareModelIcon.setFitHeight(25);
+                shareModelIcon.setFitWidth(25);
+                shareModelBtn.setGraphic(shareModelIcon);
+                shareModelBtn.setStyle("-fx-background-color: none;");
+                shareModelBtn.setOnAction(shareModelBtnClicked);
+
+                modelMeshPane.getChildren().add(shareModelBtn);
+                StackPane.setAlignment(shareModelBtn, Pos.BOTTOM_RIGHT);
+                shareModelBtn.setTranslateX(-60);
+
+            } else if (!model.get("owner_id").asString().getValue().equals(app.users.ownerId) && model.get("isShared").asBoolean().getValue()) {
+                // Current user is not the owner and the model is shared
+
+                // Show 'Shared with me' banner on models that apply
+                Text sharedBannerLabel = new Text("Shared with me");
+                sharedBannerLabel.setStyle("-fx-fill: #55b0ff; -fx-font-size: 12px; -fx-font-family: Arial; -fx-padding: 0; -fx-background-insets: 0");
+                StackPane sharedBannerPane = new StackPane(sharedBannerLabel);
+                StackPane.setAlignment(sharedBannerLabel, Pos.CENTER);
+                sharedBannerPane.setMinWidth(105);
+                sharedBannerPane.setMinHeight(30);
+                sharedBannerPane.setMaxWidth(105);
+                sharedBannerPane.setMaxHeight(30);
+                sharedBannerPane.setStyle("-fx-background-color: #171a1d; -fx-background-radius: 8 0 8 0;");
+                modelMeshPane.getChildren().add(sharedBannerPane);
+                StackPane.setAlignment(sharedBannerPane, Pos.TOP_LEFT);
+            } else {
+                // Current user is the owner and the model is not shared
+                Button shareModelBtn = new Button();
+                ImageView shareModelIcon = new ImageView("/images/share-btn.png");
+                shareModelIcon.setFitHeight(25);
+                shareModelIcon.setFitWidth(25);
+                shareModelBtn.setGraphic(shareModelIcon);
+                shareModelBtn.setStyle("-fx-background-color: none;");
+                shareModelBtn.setOnAction(shareModelBtnClicked);
+
+                modelMeshPane.getChildren().add(shareModelBtn);
+                StackPane.setAlignment(shareModelBtn, Pos.BOTTOM_RIGHT);
+                shareModelBtn.setTranslateX(-60);
+            }
+
             myModelsFlowPane.getChildren().add(modelMeshPane);
         } catch (Exception exception) {
             // Handle errors
@@ -157,7 +275,9 @@ public class myModelsController {
     EventHandler<ActionEvent> deleteModelBtnClicked = new EventHandler<ActionEvent>() {
 
         /**
-         * Deletes a model preview card from the my models view and removes the corresponding model from the database
+         * If the model is a shared model, the user removes themselves from the collaboration.
+         * Otherwise, the model is not shared and therefore is deleted from the my models view
+         * and removes the corresponding model from the database.
          *
          * @param event a JavaFX ActionEvent
          */
@@ -165,7 +285,9 @@ public class myModelsController {
         public void handle(ActionEvent event) {
             StackPane currentModel = (StackPane) ((Button) event.getSource()).getParent();
 
-            // Delete the model from the myModels view
+            BsonDocument modelDocument = app.dashboard.dbModelsList.get(app.dashboard.getDocumentIndexByModelID(app.dashboard.dbModelsList, currentModel.getId())).asDocument();
+
+            // Remove the model from the myModels view no matter what
             myModelsFlowPane.getChildren().remove(currentModel);
             app.dashboard.dbModelsList.remove(
                     app.dashboard.dbModelsList.get(
@@ -179,10 +301,15 @@ public class myModelsController {
             noModelsBtn.setVisible(false);
             myModelsScrollPane.setVisible(false);
 
-            // Delete the model from the database
-            String functionCall = String.format("ModelBox.Models.deleteCurrentUserModel('%s');", currentModel.getId());
-            app.mongoApp.eval(functionCall);
-
+            if (modelDocument.get("isShared").asBoolean().getValue() && !modelDocument.get("owner_id").asString().getValue().equals(app.users.ownerId)) {
+                // Remove this user from the collaboration
+                String functionCall = String.format("ModelBox.Models.terminateModelCollaboration('%s');", currentModel.getId());
+                app.mongoApp.eval(functionCall);
+            } else {
+                // Delete the model from the database
+                String functionCall = String.format("ModelBox.Models.deleteCurrentUserModel('%s');", currentModel.getId());
+                app.mongoApp.eval(functionCall);
+            }
         }
     };
 
@@ -196,66 +323,10 @@ public class myModelsController {
         @Override
         public void handle(ActionEvent event) {
             StackPane currentModel = (StackPane) ((Button) event.getSource()).getParent();
-            Parent previewRoot = null;
 
-
-            // Load a preview pop-up window
-            try {
-                app.viewLoader = new FXMLLoader(getClass().getResource("/views/myModels/previewPopUp.fxml"));
-                previewRoot = app.viewLoader.load();
-                app.previewPopUpView = app.viewLoader.getController();
-            } catch (Exception exception) {
-                // Handle errors
-                exception.printStackTrace();
-            }
-
-            // Load the model file from the database models list
-            int modelIndex = app.dashboard.getDocumentIndexByModelID(app.dashboard.dbModelsList, currentModel.getId());
-            BsonDocument model = app.dashboard.dbModelsList.get(modelIndex).asDocument();
-
-            int numBytes = model.get("modelFile").asBinary().getData().length;
-
-            try {
-                app.dashboard.stlImporter.read(DataURLs.builder(model.get("modelFile").asBinary().getData()).withBase64Data(true).withMediaType("model/stl").build());
-            } catch (Exception exception) {
-                // Handle errors
-                exception.printStackTrace();
-            }
-
-            // Create the model in JavaFX
-            TriangleMesh currentModelMesh = app.dashboard.stlImporter.getImport();
-            MeshView currentModelMeshView = new MeshView(currentModelMesh);
-
-            // Set the id of the previewModelSubScene to be equal to the model id
-            app.previewPopUpView.previewModelAnchorPane.setId(currentModel.getId());
-
-            // Set the modelNameText, modelTypeText, modelSizeText, and modelDateText model labels
-            String currentModelName = model.get("modelName").asString().getValue();
-            app.previewPopUpView.modelNameTextField.setText(FilenameUtils.removeExtension(currentModelName));
-            app.previewPopUpView.modelTypeText.setText(FilenameUtils.getExtension(currentModelName).toUpperCase());
-            app.previewPopUpView.modelSizeText.setText(app.models.getModelSize(numBytes));
-            String modelId = currentModel.getId();
-            app.previewPopUpView.modelDateText.setText(app.models.getModelTimestamp(modelId));
-
-            // Add the mesh to the preview pop-up group
-            Group previewModelGroup = new Group(currentModelMeshView);
-
-            // Add the group to the sub-scene for manipulation
-            app.previewPopUpView.previewModelSubScene.setRoot(previewModelGroup);
-
-            // Add the camera to the sub-scene
-            Camera camera = new PerspectiveCamera();
-            app.previewPopUpView.previewModelSubScene.setCamera(camera);
-
-            // Position the sub-scene so that it extends to the modelAnchorPane bounds
-            app.previewPopUpView.previewModelSubScene.widthProperty().bind(app.previewPopUpView.previewModelAnchorPane.widthProperty());
-            app.previewPopUpView.previewModelSubScene.heightProperty().bind(app.previewPopUpView.previewModelAnchorPane.heightProperty());
-
-            // Make the model so that it can be manipulated in 3D space
-            app.previewPopUpView.initMouseControl(previewModelGroup, app.previewPopUpView.previewModelSubScene, (Stage) myModelsAnchorPane.getScene().getWindow());
-
-            // Actually launch the preview pop-up
-            myModelsAnchorPane.getChildren().add(previewRoot);
+            // Get the collaborators for the current model and fill in the previewPopUpView
+            String functionCall = String.format("ModelBox.Models.getCurrentModelPreview('%s');", currentModel.getId());
+            app.mongoApp.eval(functionCall);
         }
     };
 
@@ -270,9 +341,44 @@ public class myModelsController {
         public void handle(ActionEvent event) {
             StackPane currentModel = (StackPane) ((Button) event.getSource()).getParent();
 
-            // Get the collaborators for the current model
+            // Get the collaborators for the current model and fill in the sharePopUpView
             String functionCall = String.format("ModelBox.Models.getCurrentModelCollaborators('%s');", currentModel.getId());
             app.mongoApp.eval(functionCall);
+        }
+    };
+
+    EventHandler<ActionEvent> compareCheckboxClicked = new EventHandler<ActionEvent>() {
+
+        /**
+         * Opens a share pop-up panel for the user to share and edit permissions for a specific model
+         *
+         * @param event a JavaFX ActionEvent
+         */
+        @Override
+        public void handle(ActionEvent event) {
+            StackPane currentModel = (StackPane) ((CheckBox) event.getSource()).getParent();
+            checkboxCount++;
+
+            if (checkboxCount == 2) {
+                // Load a share pop-up window
+                try {
+                    checkboxCount = 0;
+                    Parent shareRoot = null;
+
+                    app.viewLoader = new FXMLLoader(getClass().getResource("/views/myModels/sharePopUp.fxml"));
+                    shareRoot = app.viewLoader.load();
+                    app.sharePopUpView = app.viewLoader.getController();
+
+                    // Set the id of the shareRootAnchorPane to be equal to the model id
+                    app.sharePopUpView.shareRootAnchorPane.setId(currentModel.getId());
+
+                    // Actually launch the share pop-up
+                    app.myModelsView.myModelsAnchorPane.getChildren().add(shareRoot);
+                } catch (Exception exception) {
+                    // Handle errors
+                    exception.printStackTrace();
+                }
+            }
         }
     };
